+++ mixins/async-container/demo/async-container.html
@@ -7,9 +7,9 @@
 		<link rel="stylesheet" href="../../demo/styles.css" type="text/css">
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
-			import '../../demo/demo-page.js';
-			import '../../button/button.js';
-			import '../../loading-spinner/loading-spinner.js';
+			import '../../../components/demo/demo-page.js';
+			import '../../../components/button/button.js';
+			import '../../../components/loading-spinner/loading-spinner.js';
 			import './async-container.js';
 			import './async-item.js';
 		</script>
