+++ mixins/async-container/test/async-container-mixin.visual-diff.html
@@ -4,7 +4,7 @@
 		<link rel="stylesheet" href="../../../test/styles.css" type="text/css">
 		<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
 		<script type="module">
-			import '../../typography/typography.js';
+			import '../../../components/typography/typography.js';
 			import '../demo/async-container.js';
 			import './async-item.js';
 		</script>
